refactor components to reduce size of top level
