+++ src/components/PlayerView/Rooms/RoomContent/index.tsx
@@ -1,15 +1,8 @@
 import React from 'react'
-import { Message } from '../MessageBubble'
-import MessageInputForm from '../MessageInputForm'
-import MessageList from '../MessageList'
 import './index.css'
-
-export interface Room {
-  id: string
-  name: string
-  unreadCount: number
-  messages: Message[]
-}
+import MessageInputForm from '../../Messages/MessageInputForm'
+import MessageList from '../../Messages/MessageList'
+import { Room } from '../../../../types/wargame'
 
 const RoomContent: React.FC<{
   room: Room
